fix(index.js): 게임삭제 프롬프트에서 취소 눌렀을 때 fetch 하지 않도록 수정

+++ src/main/resources/static/src/index.js
@@ -111,6 +111,9 @@ function showBoard() {
 
 function deleteBoard(id) {
     const inputPassword = prompt('비밀번호를 입력하세요');
+    if (inputPassword == null) {
+        return;
+    }
     fetch(`/board?id=${id}`, {
         method: "DELETE",
         headers: {
